# Mythos/mythos-content-engine/frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TellTales Ink API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/generate": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Legacy Generate",
        "operationId": "legacy_generate_api_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Legacy Generate Api Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/podcasts/scripts": {
      "post": {
        "tags": [
          "podcasts"
        ],
        "summary": "Create Script",
        "operationId": "create_script_api_podcasts_scripts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PodcastGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/podcasts/generate": {
      "post": {
        "tags": [
          "podcasts"
        ],
        "summary": "Legacy Create Script",
        "operationId": "legacy_create_script_api_podcasts_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PodcastGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Legacy Create Script Api Podcasts Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/podcasts/{draft_id}": {
      "get": {
        "tags": [
          "podcasts"
        ],
        "summary": "Get Podcast",
        "operationId": "get_podcast_api_podcasts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "podcasts"
        ],
        "summary": "Update Podcast",
        "operationId": "update_podcast_api_podcasts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentGenerationRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "related_book": {
            "type": "string",
            "title": "Related Book",
            "default": ""
          },
          "related_books": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Books"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": ""
          },
          "audience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audience"
          },
          "objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objectives"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "cta": {
            "type": "string",
            "title": "Cta",
            "default": ""
          },
          "formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formats"
          },
          "visual_formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visual Formats"
          },
          "visual_style": {
            "type": "string",
            "title": "Visual Style",
            "default": ""
          },
          "image_source": {
            "type": "string",
            "title": "Image Source",
            "default": ""
          },
          "hashtag_mode": {
            "type": "string",
            "title": "Hashtag Mode",
            "default": ""
          },
          "blog_length": {
            "type": "string",
            "title": "Blog Length",
            "default": ""
          },
          "blog_format": {
            "type": "string",
            "title": "Blog Format",
            "default": ""
          },
          "structure": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Structure"
          },
          "spoiler_level": {
            "type": "string",
            "title": "Spoiler Level",
            "default": ""
          },
          "source_restrictions": {
            "type": "string",
            "title": "Source Restrictions",
            "default": ""
          },
          "length_preference": {
            "type": "string",
            "title": "Length Preference",
            "default": ""
          },
          "supporting_image": {
            "type": "string",
            "title": "Supporting Image",
            "default": ""
          },
          "image_generation": {
            "type": "string",
            "title": "Image Generation",
            "default": ""
          },
          "word_count": {
            "type": "string",
            "title": "Word Count",
            "default": ""
          },
          "seo_keyword": {
            "type": "string",
            "title": "Seo Keyword",
            "default": ""
          },
          "heading_depth": {
            "type": "string",
            "title": "Heading Depth",
            "default": ""
          },
          "metadata_controls": {
            "type": "string",
            "title": "Metadata Controls",
            "default": ""
          },
          "knowledge_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Sources"
          },
          "source_focus": {
            "type": "string",
            "title": "Source Focus",
            "default": ""
          },
          "press_timing": {
            "type": "string",
            "title": "Press Timing",
            "default": ""
          },
          "publication_date": {
            "type": "string",
            "title": "Publication Date",
            "default": ""
          },
          "byline_city": {
            "type": "string",
            "title": "Byline City",
            "default": ""
          },
          "news_angle": {
            "type": "string",
            "title": "News Angle",
            "default": ""
          },
          "auto_news_angle": {
            "type": "boolean",
            "title": "Auto News Angle",
            "default": false
          },
          "boilerplate": {
            "type": "string",
            "title": "Boilerplate",
            "default": ""
          },
          "auto_boilerplate": {
            "type": "boolean",
            "title": "Auto Boilerplate",
            "default": false
          },
          "media_contact": {
            "type": "string",
            "title": "Media Contact",
            "default": ""
          },
          "media_contact_name": {
            "type": "string",
            "title": "Media Contact Name",
            "default": ""
          },
          "media_contact_email": {
            "type": "string",
            "title": "Media Contact Email",
            "default": ""
          },
          "media_contact_phone": {
            "type": "string",
            "title": "Media Contact Phone",
            "default": ""
          },
          "media_contact_website": {
            "type": "string",
            "title": "Media Contact Website",
            "default": ""
          },
          "companion_assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Companion Assets"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "topic"
        ],
        "title": "ContentGenerationRequest"
      },
...
      "DraftUpdateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Draft"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "DraftUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PerformanceCue": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "PerformanceCue"
      },
      "PodcastDraft": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "podcast"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Draft"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          },
          "path": {
            "type": "string",
            "title": "Path",
            "default": ""
          },
          "source_path": {
            "type": "string",
            "title": "Source Path",
            "default": ""
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "show_title": {
            "type": "string",
            "title": "Show Title",
            "default": ""
          },
          "episode_title": {
            "type": "string",
            "title": "Episode Title",
            "default": ""
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/PodcastSpeaker"
            },
            "type": "array",
            "title": "Speakers"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/PodcastSegment"
            },
            "type": "array",
            "title": "Segments"
          },
          "raw_script": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Script"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "PodcastDraft"
      },
      "PodcastGenerationRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title",
            "default": ""
          },
          "show_title": {
            "type": "string",
            "title": "Show Title",
            "default": ""
          },
          "episode_title": {
            "type": "string",
            "title": "Episode Title",
            "default": ""
          },
          "episode_number": {
            "type": "string",
            "title": "Episode Number",
            "default": ""
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "default": "Spotify"
          },
          "podcast_format": {
            "type": "string",
            "title": "Podcast Format",
            "default": "roundtable discussion"
          },
          "speakers": {
            "type": "string",
            "title": "Speakers",
            "default": "3"
          },
          "speaker_roles": {
            "type": "string",
            "title": "Speaker Roles",
            "default": "Host, Co-host, Guest"
          },
          "tone": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tone"
          },
          "audience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audience"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "performance_cues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Performance Cues"
          },
          "custom_constraints": {
            "type": "string",
            "title": "Custom Constraints",
            "default": ""
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "default": "eleven_multilingual_v2"
          },
          "target_length": {
            "type": "string",
            "title": "Target Length",
            "default": "5-7 minutes"
          },
          "cta": {
            "type": "string",
            "title": "Cta",
            "default": ""
          },
          "knowledge_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Sources"
          },
          "source_focus": {
            "type": "string",
            "title": "Source Focus",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "PodcastGenerationRequest"
      },
...
      "PodcastSegment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/PodcastTurn"
            },
            "type": "array",
            "title": "Turns"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "turns"
        ],
        "title": "PodcastSegment"
      },
      "PodcastSpeaker": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PodcastSpeaker"
      },
      "PodcastTurn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "speaker_id": {
            "type": "string",
            "title": "Speaker Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "cues": {
            "items": {
              "$ref": "#/components/schemas/PerformanceCue"
            },
            "type": "array",
            "title": "Cues"
          },
          "estimated_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "speaker_id",
          "text"
        ],
        "title": "PodcastTurn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}